Rename camel case config keys to snake cases for consistency
