+++ tests/dbt/config.yml
@@ -1,3 +1,6 @@
+---
 manifest: manifest
 catalog: catalog
+docs_base_url: http://localhost
+project_source_url: http://foo.bar
 output:
