+++ tests/dbt/test_config.py
@@ -17,5 +17,7 @@ def test_yaml_config(test_root_dir):
     assert config == DbtRunConfig(
         manifest="manifest",
         catalog="catalog",
+        docs_base_url="http://localhost",
+        project_source_url="http://foo.bar",
         output=None,
     )
